Reland "[maglev] Inline GetIterator"
This is a reland of commit 2098e649b4ab11274b72dc1252543ffbcef980f5

Fix: LazyContinuationFrameScope should be only considered for
lazy deopts. We can skip when emitting an eager frame.

Before this CL we would checkpoint the eager deopt frame, so that
it wouldn't be recreated with the lazy continuation frame on top.
However, this CL wraps a lazy deopt continuation to a call to
TryBuildLoadNamedProperty, which can split in multiple blocks
in case of polymorphic feedback and need to reconstruct the
deopt frame.

I also added the regression test.

Original change's description:
> [maglev] Inline GetIterator
>
> Bug: v8:7700
> Change-Id: I8bc9ac2bc884e4f02e159aaf1a9b04b91fc4be4e
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5531447
> Commit-Queue: Victor Gomes <victorgomes@chromium.org>
> Reviewed-by: Toon Verwaest <verwaest@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#93856}

Bug: v8:7700
Change-Id: I90fe9ac7fe6be7ffb64c35aa02f1c4c63c853c92
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5537911
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93906}