[handles] Remove redundant check and obsoleted #if
In Sep 2023, a change was made to Handle::operator -> that added a
new static_assert(false). This code doesn't compile on older clang
builds, so an #if was added on __clang_version__ to bypass the new
check. Using this specific variable is fraught with peril, as it's
used by Apple to report not the version of LLVM but that of Xcode.

Additionally--and worse--this check did not compile correctly with
some Google forks of LLVM--including the one for the Android NDK--
as the specific patch to fix this clang bug was missed in the LLVM
17 merge. Apparently, this was caught by someone else and fixed in
Apr 2024, by replicating the check, replacing false in the assert.

As the underlying static_assert(false) is gone, both the condition
and the comment describing the bug are now incorrect and should be
removed. Additionally, it feels better to reorganize this function
so that it matches the code of Handle::operator *; this simplifies
the logic and avoids duplication of the static_assert's condition.

Bug: 339682307
Change-Id: I94f7f32b52032dd77ccac8c98c9eb5611ef73f95
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5541851
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93908}