[heap] Don't sweep promoted pages
Promoted pages are still iterated and zapped if needed, but we're not
building a free list and the pages won't be available for allocation.
The whole page is accounted as allocated.
This change is meant to limit the number of pages accumulated in old
space between full GCs, by accounting for promoted pages as fully
allocated (rather than only accounting a few live bytes as allocated),
and thus advancing towards the allocation limit.

Bug: 40096225
Change-Id: I95ff21efa3ebd47730430e49664fcd68612d5f29
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5401779
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94116}