[wasm][fastapi] Don't support fallback and options parameter anymore
Originally the fast API had a fallback mechanism, such that an API
function would get called a second time with the normal API calling
conventions if unsupported functionality was used by the API function
during the first call with the fast calling conventions, e.g. when the
API function was throwing an exception.

Due to recent changes, there is no unsupported functionality anymore,
and the fallback mechanism is not needed anymore. Therefore there is no
need for WebAssembly to support it.

R=jkummerow@chromium.org

Bug: 41492790
Change-Id: I34bea9a5ecb7f9687e3c10e49f1e5d6fe896c2fe
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5537972
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93939}