[cppgc] Add a flag to decommit rather than discard pooled pages
The cppgc page pool is meant to reduce system calls when freeing and
allocating memory. To avoid wasting memory, it discards memory rather
than decomitting it. While these operations may be equivalent on some
OSes, they are not on Windows, where Chromium's implementation (at
least) is using DiscardVirtualMemory(), which per MSDN "Discards the
memory contents of a range of memory pages, without decommitting the
memory."

The distinction is important, as Windows machines typically have a fixed
system-wide "commit limit" which caps committed memory, and most OOM
crashes are actually commit limit related.

This CL adds a flag to decommit the memory, rather than discarding. This
requires to recommit the memory (rather than doing nothing) when we want
to use it again, but given that discarding happens rarely (during
memory-reducing GCs), it should not impact performance. Still, in order
to assess memory, stability and performance impacts, we add a feature
flag to perform a field trial in Chrome.

Change-Id: Ieb0f46e2f48ce23c8f4ef21b0b94ee3ae405fa2e
Bug: chromium:326303884
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5541750
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93963}