[heap] Rework marking bottlenecks
Bottleneck TryMark()/PushMarkedObject() better to allow filtering on
those transitions. This is the first of more such refactorings.

The CL also removes the unused --track-retaining-path mode. The mode
recordered retaining paths as visited by the marker but was missing
the callers for ephemerons and doesn't support Oilpan. The heap
snapshot is a better alternative at this point which allows to find
retainers as well without interferring with actual marking code.

Bug: 340989496, 335479495
Change-Id: Ia78e69a596e55a1e155da13bcd246fd692e6f78b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5547837
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93982}