Never Pause Here mutes Wasm traps
In JavaScript, a breakpoint with a false condition will prevent
stopping for other reasons (usually exceptions) on the same
statement. This is the basis of the DevTools "Never Pause Here"
feature. This change brings the same behavior to WebAssembly,
so that this can be used to ignore Wasm traps.

Bug: b:341040531
Change-Id: I9f9b5d4632c8261dcbd50130bb8882728ac76474
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5544146
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Eric Leese <leese@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93966}