[wasm][jspi] Run the wasm-to-js wrapper on the central stack
Switch to the central stack at the start of the wrapper, and switch back
to the secondary stack at the end (instead of switching just around the
call to JS).
This ensures that wasm builtins (in particular conversion builtins) are
also included in this scope, and can therefore safely call other JS
builtins. Otherwise, the JS builtins can call runtime functions (such
as the slow-path allocation) without switching to the central stack
first.

R=ahaas@chromium.org

Bug: 336852356
Change-Id: I40809c16bcd9afa01a0fa2fd1658a2a5ae7c94c2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5512754
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93976}