Add adamk to src/d8/OWNERS
As an owner in include/ and a Chromium/Blink committer, I already have
responsiblity for how V8 embedding works. Since d8 is the "simplest"
V8 embedder, it seems reasonable for me to be an owner there too
(and I've hacked on it regularly over the years).

This will add a second owner in Pacific time.

Bug: 338340739
Change-Id: Ic120931b361c2cca1dba1326662cf43f0ee056a7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5548558
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Adam Klein <adamk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93984}