[deoptimizer][wasm] Support deopt inside inlined direct return call
Pseudo code example:
  function a(target) {
    return_call b(target);
  }
  function b(target) {
    call_ref(target)();
  }
  function c() {}
  function d() {}

In this case we might have type feedback for b that target == c.
When we optimize a, we inline b and speculatively inline c with a
deoptIfNot(target == c).
Now, if we call a(d), even though a is the optimized function with the
"wrong" speculation, we only construct an unoptimized output frame for
b, but not for a, as the return_call has already happened.
We still need to "throw away" the optimized code for a though as it
contains the mis-speculation.

Bug: 42204618
Change-Id: I6b6a9de6b47b97cefee610e273b63f1048e27d41
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5554494
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94024}