[heap] Speed up indirect pointer write barrier
This CL replaces the page flag checks in the indirect pointer write
barrier with directly checking the is_marking flag on the isolate.
The deferred code path is now only entered if marking is enabled,
while previously it was always entered when writing into old objects.
This should help in making the barrier faster and more cache-friendly.

We do not need page flag checks because we know that the value has
to be an ExposedTrustedObject and cannot be e.g. a read only object
like undefined.

This check is always precise except when the shared heap is enabled.
In this case is_marking may be true on client isolates for enabling
the shared marking barrier while incremental marking is not enabled
for this client isolate itself but only for the shared space/main
isolate. Should this ever become a problem, we can simply introduce a
"is_marking_local" flag that we could check instead.

Bug: 338342768
Change-Id: Ibe37352b1dc7f0bffa03b1cbfcbbd0597864c419
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5543923
Reviewed-by: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94030}