[deoptimizer][wasm] Support deopt inside inlined call_ref
This requires passing information to liftoff if the location for which
we request the LiftoffFrameDescriptionForDeopt for call_ref is actually
about a deopt point (i.e. at the beginning of the call_ref instruction
and therefore prior to the call or about an inlined location in which
case it is about the location at the call / directly after the call.

Bug: 42204618
Change-Id: Ibca128fc307049072bda5a0491780790275e8967
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5558965
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94058}