[heap] Introduce helper for bottlenecking marking worklist operation
Introduces marking helpers that can be used for full GCs and young-gen
sticky-markbit GCs. The helpers are used to bottleneck querying and
pushing which allows adding more filters in future.

- ShouldMarkObject(): Queries whether an object should be marked and
also returns the target worklist. Currently only the regular worklist
is supported but in future this can be used to signal pushing to the
on-hold worklist.
- GetLivenessMode(): Similar to ShouldMarkObject() but used for
querying whether an object should always be considered alive (e.g.
because it is in read-only space) or whether the markbit should be
used for querying its liveness.
- TryMarkAndPush(): Helper for marking and pushing an object.

Bug: 340989496, 335479495
Change-Id: I3c9c79feb1aeb843cf05a81060ce0dd85440e73a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5553775
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94054}