[heap] Clear stale slots on the stack as well
Tagged stack slots in Maglev remain live for the whole function
execution. Therefore it can happen that stale stack slots still have
pointers to objects that got left-trimmed. This leads to DCHECK
failures during GC. However, this is actually okay for Maglev because
those stack slots are not used anymore.

Ignition and Sparkplug do not have this issue because they generally
do not have pointers to FixedArrays on the stack. Turbofan does not
make stale stack slots part of the safepoint.

Main thread handles have the same issue as Maglev. Here we already
clear left-trimmed slots through a visitor before iterating these
slots for the GC. We can use the same mechanism for Maglev stack
frames as well. While this is actually only necessary for Maglev
stack frames, we currently do not have this information about stack
slots in the RootVisitor anymore. So for simplicity this CL uses this
visitor for all managed stack frames.

So far ClearStaleLeftTrimmedHandlesVisitor was a separate pass over
all slots, this CL changes this class to forward all slots to the
actual visitor. This avoids iterating the stack twice.

Bug: 40063472
Change-Id: I6fc40df7b4df5f54aab616163ec93f063f78af50
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5563330
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94069}