[wasm][fastapi] Statically pick correct function overload
Fast API calls allow function overloading, as long as the signatures of
the function either differ in the number of parameters, or in their
types. Most function overloading gets resolved at compile time, but if
two functions only differ in one parameter which is a JSArray for one
function but a TypedArray for the other function, then the function
overload has to be resolved during runtime.

So far, WebAssembly did not support function overloading. A fast API
call was possible if the first function had a matching signature,
otherwise a normal API call was used. With this CL, all functions get
checked if there signature matches, and the one function with the
matching signature gets called. However, JSArray parameters are not
allowed, to make sure that dynamic overload resolution is not necessary.

There is still a runtime check missing, i.e. if the expected parameter
is a TypedArray, there has to be a runtime check that the incoming value
is indeed a TypedArray. The code for the runtime check is missing in
this CL, and will be done in a separate CL.

R=jkummerow@chromium.org

Bug: 41492790
Change-Id: Icc50ac0a1ed3cf350e1a5cc13b20122a30d56c1a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5537974
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94251}