[heap] Move marking bit into separate byte for traced handles
This CL moves the markbit from the flags_ field into a separate
boolean field. This should not regress the size of TracedNode on
either 32- or 64-bit architectures as we still use 2 words for a
TracedNode. With this change the flags are now main thread only and
do not need use atomic accesses anymore.

This CL should also fix a data race where is_in_use() was used
non-atomically in a DCHECK while the concurrent marker set the mark
bit concurrently.

Bug: 340789840
Change-Id: I73fa6e9cc22fa3a6fc1d54922114911f9bb0db3a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5569191
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94072}