[sandbox] Manually assign IndirectPointerTag values
Previously, the IndirectPointerTags (used to ensure type-safe access to
trusted objects) were constructed simply by shifting the instance type
id to the left. This, however, is not guaranteed to result in conformant
tags: for that, it must be guaranteed that for every pair of tags T1,
T2, the result of T1 & ~T2 is nonzero. With this CL, the tags are now
manually assigned to guarantee this property.

Bug: chromium:337941142
Change-Id: I92bf7f0450b5dd885c2a556db3759782082229fa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5572948
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94106}