[ic] Extend clone IC to support different representations
When a clone IC contains a source-target pair of maps used in fast
cloning, we must ensure that both ends of the clone agree on field
representations. This CL adds a dependency using a special transition
that allows a reconfiguration of the source representation to propagate
to the target representation.

Bug: chromium:40764103
Change-Id: I4c5cb7cd461a8ed5608cc267cd040d13dfb7df4f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5505450
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94107}