[sandbox] Introduce CppHeapPointerTable and custom tagging scheme
For pointers coming from the Embedder (such as various Blink objects
that are wrapped by V8 JavaScript objects), we currently use the
ExternalPointerTable. However, that table's type-tagging scheme only
supports 70 different types, while Blink has hundreds or even thousands
of different types. Furthermore, Blink makes use of type hierarchies,
and so a type check for object A must succeed for an object B iff B is a
(direct or indirect) subclass of A. As this isn't possible with the
current scheme, all objects coming from the embedder currently use the
same type tag, which is not robust against attacks.

To address this, this CL now introduces a dedicated CppHeapPointerTable
for pointers coming from the embedder. This table uses a different type
tagging scheme based on range checks instead of bitwise ANDs. With that
scheme, the type id (a 15-bit integer) is stored in the unused parts of
the pointer. On access, the actual id is simply checked to be within the
expected range of types. This allows for 2**15 different types while
also allowing for subtype checks: a check for object A would simply
check the range for all subtypes of A.

In practice, this new scheme requires a few more instructions than the
old scheme (2-3 on Arm64), but since these are constant-time operations,
their overhead appears to be mostly negligible in practice.

Bug: chromium:40940619
Change-Id: I39049c523832b55a4ad7b3ae09f5b604306fb3c1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5563324
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94112}