[sandbox] Fix false-positive ASan crash
When loading a pointer from the TrustedPointerTable with the wrong type,
the returned pointer will point to a non-canonical address and will
always crash in a safe way when dereferenced. However, when ASan is
enabled, any dereference is preceeded by a shadow memory access. In that
case, the math used to compute the address of the shadow memory may end
up producing a canonical (but unmapped) address, thereby leading to a
segfault that appears to be unsafe. This CL fixes that by using an
explicit CHECK in ASan-enabled builds.

Bug: chromium:342451736
Change-Id: I997d9230641402cb5749587a75f846f7d5de7806
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5573734
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94125}