[sandbox] Tweak ReadCppHeapPointerField
Previously, we required an explicit check for the null handle, in which
case we would return nullptr. With this CL we instead guarantee that an
access to the 0th entry of the table always results in nullptr being
returned. As such, the check for the null handle is no longer needed. On
Arm64 this is a net improvement as returning nullptr for the 0th entry
is possible without any additional instructions. On x64, this change
should be neutral as it effectively replaces one conditional branch with
another one.

Bug: chromium:40940619
Change-Id: Ieeba3f041cefc8d080f5b22f972f591b503cccd2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5576554
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94155}