[sandbox] Introduce v8::Object::Unwrap variant with tag range
The new pointer tagging scheme introduced in https://crrev.com/c/5563324
makes use of range-based type checks and therefore requires a tag range
when loading a pointer from the pointer table. This CL introduces the
necessary variant of v8::Object::Unwrap to support this. As the tag
ranges do not necessarily need to be compile-time constants, this CL
also switches the relevant routines to take the tag range as regular
parameter, not as template parameter. In the (common) case where the tag
range is known at compile-time, the compiler should be able to optimize
the code in the same way.

Bug: 40940619, 342446331
Change-Id: I76083d4fb0b92bc8ebc2b17e3ea271128e0047fc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5580038
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94147}