[wasm] Support calling imports with the generic export wrapper
So far the generic wrapper was only used for calling internal wasm
functions, which allowed it to make the assumption that the implicit
parameter was a WasmTrustedInstanceData object.
Add support for passing a WasmApiFunctionRef, which enables using it for
calling re-exported imports, and also fixes a correctness issue with the
JSPI variant of the wrapper.

R=ahaas@chromium.org

Fixed: 342415789
Change-Id: I6c49606ebb5d4406d6ed878b07fff3c8b5b2415c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5575093
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94161}