[sandbox] Refactor base::AbortMode
Currently, for sandbox fuzzing and sandbox testing mode, we use the
"soft" abort mode. This causes DCHECKs failures to be ignored entirely,
and CHECK failures, UNREACHABLEs, etc. to terminate the process through
exit(-1) as these are considered harmless crashes for the sandbox. The
issue with that is that for writing regression tests, exit(-1) does not
count as "passing" a test. On the other hand, we also do not want to
use exit(0) (i.e. indicate success) in sandbox fuzzing mode, as the
fuzzer should treat such cases as failed executions and discard them.
Therefore, this CL now removes the soft abort mode entirely and replaces
it with a AbortMode::kExitWithSuccessAndIgnoreDcheckFailures and
AbortMode::kExitWithFailureAndIgnoreDcheckFailures, which are quite
verbose, but at least fairly descriptive.

Bug: 42202821
Change-Id: I9d8f887a2c0cad051e2b495a61b8508b41043d2e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5580879
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94172}