[heap] Actually wait for page to be swept
`EnsurePageIsSwept` tries to steal a page from the sweeper's worklist if
the sweeper hasn't started sweeping it yet. In case the status says that
the page is not yet swept, but the sweeper starts handling it after the
status is read and before the page is stolen, stealing it will fail and
the method will return while the page is not yet swept.

Bug: 343484151
Change-Id: I14aea4e1ce1db3d4698d24d145f3979493e79dbb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5587857
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94173}