[sandbox] Fix handling of JSArrayBuffers during serialization
When serializing a JSArrayBuffer in the snapshot serializer, we
temporarily clear the extension, then re-set it afterwards. As the
extension is now a ExternalPointerTable::ManagedResource, this operation
has side-effects: the clearing will mark the extension as no-longer
managed by the table, and the re-setting should again mark it as such.
However, that last step was missing previously. This CL fixes that.

Fixed: 343257586
Change-Id: I0889f061835687803881d527754c61b83093ca0f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5587858
Auto-Submit: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94194}