[sandbox] Add SBXCHECK when reading a wasm::ValueKind from the heap
The ValueKind is read from inside the sandbox (e.g. a WasmTableObject)
and so is considered untrusted. Currently, when for example obtaining
the name of a ValueKind, the kind is used as an index into a global
table. As such, a corrupted ValueKind can lead to out-of-bounds reads
outside the sandbox. While these are not technically sandbox violations,
we should still avoid them to keep the fuzzers happy. This CL achieves
that by adding a SBXCHECK when reading the ValueKind from the heap.

Bug: chromium:340167178
Change-Id: I82c3783c7faefc7e873c2f3f608bbd2db36d8484
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5575087
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94190}