[sandbox] Use parameter count from BytecodeArray in interpreter
Previously, when suspending or resuming generators in the interpreter,
the formal parameter count from the SharedFunctionInfo was used to
determine how many argument values to read from the stack. This,
however, is not compatible with the sandbox as the SharedFunctionInfo
lives inside the sandbox, and therefore must be considered untrusted.
With this CL, the interpreter now instead uses the parameter count from
the (trusted) BytecodeArray object.

Bug: 337547182, 40931165
Change-Id: Ifb04f84d340633818430952ecff4fefebaaa2501
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5587863
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94198}