[map] Fix map updater failing to connect split map tree
When the MapUpdater in AddMissingTransitions creates a fresh map tree it
searches for the first map that can be shared. Starting from that split_map it creates a new sub-tree. The descriptors of the split_map
are updated *after* AddMissingTransitions. If the split_map happens to
to be a root map, then this inconsistent state will look like it is a
detached map. This causes InstallDescriptor not to insert any
transition and thus we end up creating just detached maps.

Fixed: 344654897
Change-Id: Ie7dbc94f256373c7f6f70489765d35aed9768c5e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5596271
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94242}