[map] Don't generalize fields on overflowing prototype transitions
When we can't connect a prototype transition because the prototype
transitions overflowed, we would generalize all fields. These lines
were  introduced to work around the bug that was ultimatively fixed in
https://crrev.com/c/5575086.

While it might be beneficial to still generalize fields in this case,
unfortunately we can't (without bigger changes to the MapUpdater).
The reason is that otherwise the MapUpdater creates transitions where
the properties of the root map (if it has any) go from the generalized
root maps state to a more specialized map (using old_descriptors as a
template).

For now, just remove the generalization.

Drive-By: * Also ensure a slack tracking step is taken if the new map
            is not connected due to overflow.
          * Some cleanup.

Bug: 345052532
Change-Id: Ic5843b560828f347fe6962877500ff8923300470
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5600347
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94278}