[deoptimizer][wasm] Don't delete processed feedback, mark it outdated
Deleting the cached feedback can cause races with background
compilation jobs that require this feedback to be present.
Instead we only mark it as outdated. There is still a race where an
enqueued background compilation job can read the feedback after the
deopt but before the cached feedback was updated. However, in this case
it only wastes a background compilation that will likely be running into
the deopt again.

In most cases there won't be other compile jobs for the deopted
functions enqueued or they will not be processed before the main thread
reached the TriggerTierUp call in CallRefIC which will then get an
exclusive lock and reprocess the feedback.

This CL also disables --jit-fuzzing in most deopt test cases as they
check assertFalse(%IsTurboFanFunction(wasm.main)) after the deopt which
cannot be guaranteed in this mode.

Fixed: 345403263
Bug: 42204618
Change-Id: Ice7fed0fc6c4d8f60b3e82f4edf0e43245aa6d4e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5604483
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94298}