[turboshaft] Fix VariableReducer bug with loop phis
When a Variable is alive before a loop, then the VariableReducer has
to conservatively emit a PendingLoopPhi at the begining of the loop,
and change it into a Phi afterwards.

However, this was broken for variables that did not change in the
loop, leading to the initial PendingLoopPhi never being replaced by a
regular Phi.

Bug: chromium:42202729
Change-Id: Ib11859b957f940fcdb0278f6f2ab62a24b8beeac
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5596548
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94301}