[logging] Introduce V8::SetFatalErrorHandler to hook into V8_Fatal()
This introduces a new error handler, CheckFatalErrorHandler, with the same signature as the DcheckErrorHandler callback. This allows an embedding application to hook into V8_Fatal in order to report the error, or insert any special process teardown logic as needed by the application. If this method is not set, or does not itself terminate the process, the normal logging and termination procedure is used.

BUG=v8:14671

Change-Id: Iea2003d1296db6cd0d3125e35e7db0b5e60c5d8e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5369342
Commit-Queue: Caitlin Potter <caitp@igalia.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94318}