[tagged] Introduce a Union type
Add a variadic templated C++ union type similar to torque's unions,
which encapsulates a value that is one of a limited set of V8 types.
Tagged pointers to a Union type can be cast (casted?) to one of the
types in the Union in order to access that value, or indeed to a
sub-Union if necessary.

Unions are connected to their types using the `is_subtype` type_trait,
which is substantially expanded to keep working (in trivial cases)
when its types are only forward declared. The sub-type relation for
a Union and another type defines a partial order, as one would expect,
where:

  1. Union<T...> is a subtype of U if all T are subtypes of U
  2. U is a subtype of Union<T...> if for any T, U is a subtype of T

This type is used, initially, as follows:

  1. Torque union types now generate a Union<...> instead of deferring
     to the parent class
  2. C++ defines typedefs for Number, Numeric, JSPrimitive, JSAny, and
     JSCallable. Notably, Object is not (yet?) a Union of Smi and
     HeapObject, as we need to decide what to do with the static
     methods on the Object class.
  3. Union<...> subsumes the UnionT<T,U> types in TNodes and
     Turboshaft.

The bulk of the changes in this patch are propagations of types
necessary to satisfy newly strict requirements of torque-generated
functions (in particular, setters). There is also a partial progress on
making Object::NumberValue take a Number rather than an Object, as a
proof-of-concept for Number.

Change-Id: I87868f65eba43a3b5cbda6c0bf5e16e62a8225c7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5600543
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94373}