[wasm] Pass missing module for wasm-to-js tierup of a WasmFuncRef
This tier-up path assumed that we don't have an instance. However this
is only true when the WasmFuncRef was created for a WebAssembly.Function
object.

For call refs to an internal wasm function, we do have an instance, and
the signature may contain indexed ref types, which requires that we pass
the module to the wrapper graph builder.

This will also allow some simplifications when we enable the generic
wrapper for re-exported imports again.

R=ahaas@chromium.org

Bug: 343917751
Change-Id: I23d69f0db98edd43494bcd08763b4bef9c7f0e0c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5621892
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94376}