[wasm] Do not generate import-specific export wrappers
The signature-specific export wrappers load the target and ref from the
import dispatch table or from the function data depending on whether the
target is a re-exported import or an internal function. But the target
and ref from the function data is more generic and also works for
re-exported imports.

In fact wrapper tier-up may accidentally replace an export wrapper
compiled for an import with an export wrapper compiled for an internal
function, but we did not notice it because the latter still works.

However the other way around does not work. The wrapper compiled for a
re-exported import uses the dispatch table, which breaks if the target
is not an import. This happens if we try to enable the generic wrapper
for re-exported imports.

Remove the import-specific path entirely since the other one is more
generic, which allows us to use the generic wrapper for re-exported
imports and will reduce the number of wrappers that we need to compile.

R=jkummerow@chromium.org

Bug: 343772336
Change-Id: I54446f38971167a5128521027939707ac20d1a86
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5604484
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94380}