Add a HeapProfiler function for persisting temporary strings
Embedders which implement NameProvider may want to specify formatted
strings as names for objects, rather than static strings. Doing so is
somewhat difficult, because the embedder must manage the lifetime of any
allocated string and ensure it stays alive until V8 is done generating
the snapshot. This change adds a new function so that embedders can
easily transfer a string into the HeapProfiler while it is taking a
snapshot.

Bug: 343341610
Change-Id: I75d0ff37367b9f30c3854c9c807e33d160a10fc4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5599955
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94444}