[heap] Update allocation limits after loading
During loading we avoid GCs and accumulate allocated objects. As a
result, when starting an incremental GC after loading, the allocation
limits may be already overshot by a large margin, which would result
in immediate GC finalization. This essentially results in a fully
atomic/parallel GC which is expensive.

Fix by updating the allocation limits when starting incremental marking
for the first time after loading if the allocation limits are already
reached.

Note that we don't always get a signal to switch RAIL mode and sometimes
we leave the loading mode due to either a time limit or allocation
limits. This complicates the situation since we can't just rely to RAIL
mode changes and instead need e.g. a dedicated flag checked explicitly
(as used in this CL).

Bug: 343263980, 346498599
Change-Id: Ifc96c2ba9927c3fb7b372efea798a1b40bfc6a78
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5614370
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94436}