[deoptimizer][wasm] FrameState / deopt info for nested call_indirect
Turboshaft needs to collect FrameState data for any inlined function
which contains a deopt node (already done in a prior CL, independent of
whether it's a call_indirect or not).
In case the deopt triggers, liftoff also needs to collect the deopt
info for the deoptimizer to reconstruct the inlined frame (as well as
the outer one).

Bug: 42204618
Change-Id: I9a948ffa8ba507a0f19e46802d1be66fd61aa64d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5615410
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94399}