[wasm] Use designated initializers in some places
This uses designated initializers for some struct constructions. This
replaces the existing comments and allows to skip fields which will be
initialized later or which can be default-initialized.

All fields that we want to skip in the designated initializers need a
default value now.

R=mliedtke@chromium.org

Change-Id: I628c9b3edccf7c10e6fb767f265a657ef9447b24
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5625252
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94457}