Use a faster hash table when producing a heap snapshot
When generating a heap snapshot, much of the time is spent performing
lookups in the entries_map_. We can greatly reduce this cost by using a
flat hash table. Furthermore, we can avoid a second lookup in cases
where the entry does not yet exist and must be inserted. With both of
those changes, I measured a reduction of snapshot generation time from
5.9 seconds to 4.5 seconds on a Google Docs page.

Bug: 42203857
Change-Id: I49997f375dff691c41b1ed93404d02fdbefe658c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5626035
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94443}